Moved dataset type to Dataset.schema
