+++ docetl/dataset.py
@@ -1,9 +1,10 @@
 import os
 from concurrent.futures import ThreadPoolExecutor, as_completed
-from typing import Any, Callable, Dict, List, Optional, Union
+from typing import Any, Callable, Dict, List, Optional, Union, Literal
+from pydantic import BaseModel
 
 from docetl.parsing_tools import get_parser, get_parsing_tools
-from docetl.schemas import ParsingTool
+from docetl.base_schemas import ParsingTool
 
 
 def create_parsing_tool_map(
@@ -39,6 +40,40 @@ class Dataset:
         user_defined_parsing_tool_map (Dict[str, ParsingTool]): A map of user-defined parsing tools.
     """
 
+    class schema(BaseModel):
+        """
+        Represents a dataset configuration in the pipeline.
+
+        Attributes:
+            type (str): The type of the dataset. Must be either 'file' or 'memory'.
+            path (str): The path to the dataset file or the in-memory data, depending on the type.
+            source (str): The source of the dataset. Currently, only 'local' is supported. Defaults to 'local'.
+            parsing (Optional[List[Dict[str, str]]]): A list of parsing tools to apply to the data. Each parsing tool
+                                                      is represented by a dictionary with 'input_key', 'function', and
+                                                      'output_key' keys. Defaults to None.
+
+        Example:
+            ```yaml
+            datasets:
+              my_dataset:
+                type: file
+                path: input.json
+                parsing:
+                  - input_key: file_path
+                    function: txt_to_string
+                    output_key: content
+            ```
+
+        Note:
+            The parsing tools are applied in the order they are listed. Each parsing tool takes the output
+            of the previous tool as its input, allowing for chained processing of the data.
+        """
+
+        type: Literal["file", "memory"]
+        path: str
+        source: str = "local"
+        parsing: Optional[List[Dict[str, str]]] = None
+        
     def __init__(
         self,
         runner,
